requirements.txt
# Core Airflow packages
apache-airflow==2.7.1
apache-airflow-providers-snowflake==4.2.0
apache-airflow-providers-common-sql==1.8.1

# Database and ORM
sqlalchemy>=1.4.0,<2.1
psycopg2-binary>=2.9.0
pyodbc>=4.0.0

# Data processing
pandas>=1.3.0
numpy>=1.21.0

# API and web requests
requests>=2.27.0
urllib3>=1.26.0

# Snowflake specific
snowflake-sqlalchemy>=1.4.0
snowflake-connector-python>=3.0.0

# Data visualization (optional)
matplotlib>=3.5.0
plotly>=5.0.0

# Big data processing (optional)
pyspark>=3.3.0

# Development and testing (optional)
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0

# Additional utilities
python-dotenv>=0.19.0
